Cast frametime to int

+++ backend/opencv_backend.py
@@ -16,7 +16,7 @@ class OpenCVBackend(Backend):
 
         framerate = 60
         # Time between video frames when running on OpenCV, in milliseconds
-        self.frametime = 1000 / framerate
+        self.frametime = int(1000 / framerate)
 
         self.base_canvas = Canvas(self.width, self.height)
         self.data_canvas = Canvas(self.width, self.height)
